+++ pkg/app/mutate_app.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/Masterminds/semver/v3"
+	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/apiextensions/v3/pkg/clientset/versioned"
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
@@ -77,12 +78,12 @@ func (m *Mutator) Mutate(request *v1beta1.AdmissionRequest) ([]mutator.PatchOper
 
 	// We check the deletion timestamp because app CRs may be deleted by
 	// deleting the namespace they belong to.
-	if !appNewCR.DeletionTimestamp.IsZero() {
+	if request.Operation == v1beta1.Update && !appNewCR.DeletionTimestamp.IsZero() {
 		m.logger.Debugf(ctx, "admitted deletion of app %#q in namespace %#q", appNewCR.Name, appNewCR.Namespace)
 		return nil, nil
 	}
 
-	result, err := m.MutateApp(ctx, *appNewCR)
+	result, err := m.MutateApp(ctx, *appNewCR, request.Operation)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -92,7 +93,11 @@ func (m *Mutator) Mutate(request *v1beta1.AdmissionRequest) ([]mutator.PatchOper
 	return result, nil
 }
 
-func (m *Mutator) MutateApp(ctx context.Context, app v1alpha1.App) ([]mutator.PatchOperation, error) {
+func (m *Mutator) MutateApp(ctx context.Context, app v1alpha1.App, operation v1beta1.Operation) ([]mutator.PatchOperation, error) {
+	if operation == v1beta1.Connect {
+		return nil, nil
+	}
+
 	var err error
 	var result []mutator.PatchOperation
 
@@ -123,6 +128,16 @@ func (m *Mutator) MutateApp(ctx context.Context, app v1alpha1.App) ([]mutator.Pa
 	if key.VersionLabel(app) != uniqueAppCRVersion && ver.Major() < 3 {
 		m.logger.Debugf(ctx, "skipping mutation of app %#q in namespace %#q due to version label %#q", app.Name, app.Namespace, appVersionLabel)
 		return nil, nil
+	}
+
+	if key.VersionLabel(app) == uniqueAppCRVersion {
+		managementClusterAppPatches, err := m.mutateManagementClusterPauseAnnotation(ctx, app, operation, appVersionLabel)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+		if len(managementClusterAppPatches) > 0 {
+			result = append(result, managementClusterAppPatches...)
+		}
 	}
 
 	labelPatches, err := m.mutateLabels(ctx, app, appVersionLabel)
@@ -219,6 +234,10 @@ func (m *Mutator) mutateKubeConfig(ctx context.Context, app v1alpha1.App) ([]mut
 func (m *Mutator) mutateLabels(ctx context.Context, app v1alpha1.App, appVersionLabel string) ([]mutator.PatchOperation, error) {
 	var result []mutator.PatchOperation
 
+	if len(app.Labels) == 0 {
+		result = append(result, mutator.PatchAdd("/metadata/labels", map[string]string{}))
+	}
+
 	// Set app label if there is no app label present.
 	if key.AppKubernetesNameLabel(app) == "" && key.AppLabel(app) == "" {
 		result = append(result, mutator.PatchAdd(fmt.Sprintf("/metadata/labels/%s", replaceToEscape(label.AppKubernetesName)), key.AppName(app)))
@@ -228,9 +247,28 @@ func (m *Mutator) mutateLabels(ctx context.Context, app v1alpha1.App, appVersion
 		result = append(result, mutator.PatchAdd(fmt.Sprintf("/metadata/labels/%s", replaceToEscape(label.AppOperatorVersion)), appVersionLabel))
 	}
 
-	if len(app.Labels) == 0 {
-		root := mutator.PatchAdd("/metadata/labels", map[string]string{})
-		result = append([]mutator.PatchOperation{root}, result...)
+	return result, nil
+}
+
+func (m *Mutator) mutateManagementClusterPauseAnnotation(ctx context.Context, app v1alpha1.App, operation v1beta1.Operation, appVersionLabel string) ([]mutator.PatchOperation, error) {
+	var result []mutator.PatchOperation
+
+	// We don't want to re-add pause annotation on already created objects
+	// (i.e. UPDATE events).
+	if operation != v1beta1.Create {
+		return nil, nil
+	}
+
+	if len(app.Annotations) == 0 {
+		result = append(result, mutator.PatchAdd("/metadata/annotations", map[string]string{}))
+	}
+
+	v, ok := app.Annotations[annotation.AppOperatorPaused]
+	if !ok {
+		result = append(result, mutator.PatchAdd(fmt.Sprintf("/metadata/annotations/%s", replaceToEscape(annotation.AppOperatorPaused)), "true"))
+	}
+	if ok && v != "true" {
+		result = append(result, mutator.PatchReplace(fmt.Sprintf("/metadata/annotations/%s", replaceToEscape(annotation.AppOperatorPaused)), "true"))
 	}
 
 	return result, nil